requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-openai
sentence-transformers
numpy
python-dotenv
python-multipart
pypdf
python-docx
pydantic
